+++ src/test/converter2/issues/gh2044/index.js
@@ -1,0 +1,18 @@
+export { other } from "./other";
+
+/** @typedef {import("./other").Foo} Foo */
+/** @typedef {import("./other").Foo} RenamedFoo */
+
+/**
+ * @typedef {import("./other").Generic<T>} Generic
+ * @template T
+ */
+
+/**
+ * @typedef {import("./other").Generic<U>} RenamedGeneric
+ * @template {string} U
+ */
+
+/**
+ * @typedef {import("./other").Generic<string>} NonGeneric
+ */